requirements.txt
azure-ai-vision-imageanalysis
openai